+++ lesson_1/triangle_area.rb
@@ -1,8 +1,8 @@
 puts 'Введите основание треугольника'
-basic = gets.chomp.to_i
+basic = gets.chomp.to_f
 
 puts 'Введите высоту треугольника'
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 area = 0.5 * basic * height
 puts "Площадь треугольника #{area}"
